Rearrange experimental implementations
