--- Assets/YggdrAshill.Ragnarok.Unity/Runtime/Experimental.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2fcc5f8a0c9fc5e45a099885a0cb78a0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
