--- Assets/YggdrAshill.Ragnarok.Unity/Runtime/Experimental/MonoEntryPoint.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 707afaa2b7e349ad8c312dfa95e9cc26
-timeCreated: 1695533013
