--- Assets/YggdrAshill.Ragnarok.Unity/Runtime/Experimental/ScriptableEntryPoint.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: b8ebb38713f44b0aa3787985b36f434b
-timeCreated: 1695539206
