+++ Assets/YggdrAshill.Ragnarok.Unity/Runtime/Lifecycle/ExplicitMonoEntryPoint.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 
-namespace YggdrAshill.Ragnarok.Experimental
+namespace YggdrAshill.Ragnarok
 {
     public sealed class ExplicitMonoEntryPoint : MonoEntryPoint
     {
