+++ Assets/YggdrAshill.Ragnarok.Unity/Runtime/Lifecycle/GameObjectLifecycle.cs
@@ -1,5 +1,4 @@
 #nullable enable
-using YggdrAshill.Ragnarok.Experimental;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -80,17 +79,9 @@ namespace YggdrAshill.Ragnarok
         [SerializeField] private MonoInstallation[] monoInstallationList = Array.Empty<MonoInstallation>();
         private IEnumerable<IInstallation> MonoInstallationList => monoInstallationList;
 
-        [SerializeField] private ScriptableEntryPoint[] scriptableEntryPointList = Array.Empty<ScriptableEntryPoint>();
-        [SerializeField] private MonoEntryPoint[] monoEntryPointList = Array.Empty<MonoEntryPoint>();
         protected override IEnumerable<IInstallation> GetInstallationList()
         {
-            var scriptableEntryPointInstallationList 
-                = scriptableEntryPointList.Select(entryPoint => entryPoint.Installation);
-            var monoEntryPointInstallationList
-                = monoEntryPointList.Select(entryPoint => entryPoint.Installation);
-
-            return ScriptableInstallationList.Concat(MonoInstallationList)
-                .Concat(scriptableEntryPointInstallationList).Concat(monoEntryPointInstallationList);
+            return ScriptableInstallationList.Concat(MonoInstallationList);
         }
     }
 }
