--- Assets/YggdrAshill.Ragnarok.Unity/Runtime/Lifecycle/IEntryPoint.cs
@@ -1,11 +1,0 @@
-#nullable enable
-using System;
-
-namespace YggdrAshill.Ragnarok
-{
-    [Obsolete("Use IInstallation directly.")]
-    public interface IEntryPoint
-    {
-        IInstallation Installation { get; }
-    }
-}
