--- Assets/YggdrAshill.Ragnarok.Unity/Runtime/Lifecycle/IEntryPoint.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: e79a00dcd9fd4ab9976b5e9d5936e1df
-timeCreated: 1685105182
