+++ Assets/YggdrAshill.Ragnarok.Unity/Runtime/Lifecycle/MonoEntryPoint.cs.meta
@@ -1,3 +1,3 @@
 fileFormatVersion: 2
-guid: 5b3ddaa7fce94f5284f4e397e93d0d73
-timeCreated: 1680318310
+guid: 707afaa2b7e349ad8c312dfa95e9cc26
+timeCreated: 1695533013
