+++ Assets/YggdrAshill.Ragnarok.Unity/Runtime/Lifecycle/MonoInstallation.cs
@@ -1,10 +1,10 @@
 #nullable enable
 using UnityEngine;
 
-namespace YggdrAshill.Ragnarok.Experimental
+namespace YggdrAshill.Ragnarok
 {
     // TODO: add document comments.
-    public abstract class ScriptableInstallation : ScriptableObject, IInstallation
+    public abstract class MonoInstallation : MonoBehaviour, IInstallation
     {
         public abstract void Install(IObjectContainer container);
     }
