+++ Assets/YggdrAshill.Ragnarok.Unity/Runtime/Lifecycle/ScriptableEntryPoint.cs.meta
@@ -1,3 +1,3 @@
 fileFormatVersion: 2
-guid: 18f72a8b54a34c668684a28cd6188f3c
-timeCreated: 1680318295
+guid: b8ebb38713f44b0aa3787985b36f434b
+timeCreated: 1695539206
