+++ Assets/YggdrAshill.Ragnarok.Unity/Runtime/Lifecycle/ScriptableInstallation.cs
@@ -1,10 +1,10 @@
 #nullable enable
 using UnityEngine;
 
-namespace YggdrAshill.Ragnarok.Experimental
+namespace YggdrAshill.Ragnarok
 {
     // TODO: add document comments.
-    public abstract class MonoInstallation : MonoBehaviour, IInstallation
+    public abstract class ScriptableInstallation : ScriptableObject, IInstallation
     {
         public abstract void Install(IObjectContainer container);
     }
