--- Assets/YggdrAshill.Ragnarok.Unity/Specification/Experimental.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b6daea3ba050d424d8d2c30d6c4a27ba
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
