+++ Assets/YggdrAshill.Ragnarok.Unity/Specification/ImplicitMonoEntryPointInstallation.cs
@@ -1,5 +1,4 @@
 #nullable enable
-using YggdrAshill.Ragnarok.Experimental;
 using UnityEngine;
 
 namespace YggdrAshill.Ragnarok.Specification
