+++ Assets/YggdrAshill.Ragnarok.Unity/Specification/ImplicitMonoEntryPointSpecification.cs
@@ -1,5 +1,4 @@
 #nullable enable
-using YggdrAshill.Ragnarok.Experimental;
 using System.Collections.Generic;
 
 namespace YggdrAshill.Ragnarok.Specification
