+++ Assets/YggdrAshill.Ragnarok.Unity/Specification/ImplicitMonoEntryPointSpecification.unity
