+++ Assets/YggdrAshill.Ragnarok.Unity/Specification/Scripts/SenderInstallation.cs
@@ -1,5 +1,4 @@
 ﻿#nullable enable
-using YggdrAshill.Ragnarok.Experimental;
 using System;
 using UnityEngine;
